# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Product Service API",
    "version": "1.0.0",
    "description": "Product microservice avec gestion modulaire des variants (color, size, capacity, weight) et pricing dynamique"
  },
...
  "paths": {
    "/api/v1/products": {
      "get": {
        "summary": "Get all products",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "merchant_id",
            "in": "query",
            "schema": { "type": "string" }
          },
          { "name": "store_id", "in": "query", "schema": { "type": "string" } },
          {
            "name": "category_id",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "min_price",
            "in": "query",
            "schema": { "type": "number" },
            "description": "minimum Price"
          },
          {
            "name": "max_price",
            "in": "query",
            "schema": { "type": "number" },
            "description": "maximum Price"
          },
          {
            "name": "colors",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Comma-separated list of colors"
          },
          {
            "name": "sizes",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Comma-separated list of sizes"
          },
          { "name": "weight", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "post": {
        "summary": "Create product",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "price"],
                "properties": {
                  "name": { "type": "string", "example": "Television" },
                  "price": {
                    "type": "number",
                    "format": "float",
                    "example": 4000000
                  },
                  "merchant_id": { "type": "string", "example": "AB37" },
                  "store_id": { "type": "string", "example": "OM21" },
                  "category_id": { "type": "string" },
                  "category_name": { "type": "string" },
                  "description": { "type": "string" },
                  "quantity": { "type": "integer", "default": 0 },
                  "images": {
                    "type": "array",
                    "items": { "type": "string", "format": "binary" },
                    "maxItems": 5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Données invalides" }
        }
      }
    },
...
    "/api/v1/products/{id}": {
      "get": {
        "summary": "Get product by ID",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": { "200": { "description": "Success" } }
      },
      "put": {
        "summary": "Update a product",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "price": { "type": "number" },
                  "description": { "type": "string" },
                  "status": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated" } }
      },
      "delete": {
        "summary": "Delete a Product",
        "tags": ["Products"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID unique du produit",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Produit supprimé avec succès" },
          "404": { "description": "Produit non trouvé" }
        }
      }
    },
...
    "/api/v1/variants/{variantId}": {
      "get": {
        "summary": "Get variant by ID",
        "tags": ["Variants"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "variantId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "Variant ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Variant details with price breakdown",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/ProductVariant" }
                  }
                }
              }
            }
          },
          "404": { "description": "Variant not found" }
        }
      },
      "put": {
        "summary": "Update variant (including price modifiers)",
        "tags": ["Variants"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "variantId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "Variant ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  { "$ref": "#/components/schemas/ProductVariantInput" },
                  {
                    "type": "object",
                    "properties": {
                      "overridePrice": {
                        "type": "number",
                        "description": "Manual price override (skips calculation)"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Variant updated with recalculated price",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/ProductVariant" }
                  }
                }
              }
            }
          },
          "404": { "description": "Variant not found" }
        }
      },
      "delete": {
        "summary": "Delete variant",
        "tags": ["Variants"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "variantId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "Variant ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Variant deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Cannot delete default variant" },
          "404": { "description": "Variant not found" }
        }
      }
    },
...
    "/api/v1/categories": {
      "get": {
        "summary": "Category List",
        "tags": ["Categories"],
        "responses": {
          "200": { "description": "Succès" }
        }
      },
      "post": {
        "summary": "Create a Category",
        "tags": ["Categories"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "example": "Électronique" },
                  "description": {
                    "type": "string",
                    "example": "Electronic devices and gadgets"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Category created" }
        }
      }
    },
    "/api/v1/categories/{id}": {
      "get": {
        "summary": "Detail of a category",
        "tags": ["Categories"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Succès" }
        }
      },
      "put": {
        "summary": "Modify a category",
        "tags": ["Categories"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "status": { "type": "string", "enum": ["active", "inactive"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Update done successfully" }
        }
      },
      "delete": {
        "summary": "Delete a category",
        "tags": ["Categories"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Category deleted" }
        }
      }
    },
...
    "/api/v1/products/{id}/images": {
      "post": {
        "tags": ["Images"],
        "summary": "Upload product Image",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "is_main": {
                    "type": "boolean",
                    "default": false
                  }
                },
                "required": ["images"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Image added" },
          "404": { "description": "Product unfindable" }
        }
      },
      "get": {
        "tags": ["Images"],
        "summary": "See product images",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Image list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ProductImage" }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{id}/images/{imageId}": {
      "delete": {
        "tags": ["Images"],
        "summary": "Delete an image",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "imageId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Image deleted" },
          "404": { "description": "Image unfindable" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductVariant": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "product_id": { "type": "string", "format": "uuid" },
          "sku": { "type": "string" },
          "barcode": { "type": "string" },
          "attributes": {
            "type": "object",
            "description": "Complete variant attributes with pricing info",
            "properties": {
              "color": {
                "type": "object",
                "properties": {
                  "value": { "type": "string", "example": "#FF0000" },
                  "displayValue": { "type": "string", "example": "Rouge Profond" },
                  "type": { "type": "string", "enum": ["hex", "text", "image"] },
                  "priceImpact": { "type": "number", "example": 50 },
                  "metadata": { "type": "object" }
                }
              },
              "size": {
                "type": "object",
                "properties": {
                  "value": { "type": "string", "example": "XL" },
                  "priceImpact": { "type": "number", "example": 10 }
                }
              },
              "capacity": {
                "type": "object",
                "properties": {
                  "value": { "type": "string", "example": "256GB" },
                  "displayValue": { "type": "string", "example": "256 Go" },
                  "basePrice": { "type": "number", "example": 1199 }
                }
              },
              "weight": {
                "type": "object",
                "properties": {
                  "value": { "type": "number", "example": 0.5 },
                  "label": { "type": "string", "example": "500g" },
                  "priceImpact": { "type": "number", "example": 5 }
                }
              }
            }
          },
          "color": { "type": "string", "description": "Flat color value from attributes.color.value" },
          "size": { "type": "string", "description": "Flat size value from attributes.size.value or attributes.capacity.value" },
          "base_price": { "type": "number", "description": "Product base price at creation" },
          "calculated_price": { "type": "number", "description": "Auto-calculated: base + color + size + weight impacts" },
          "override_price": { "type": "number", "nullable": true, "description": "Manual price override" },
          "final_price": { "type": "number", "description": "Effective price (override || calculated)" },
          "price_breakdown": {
            "type": "object",
            "properties": {
              "base": { "type": "number" },
              "color_modifier": { "type": "number" },
              "size_modifier": { "type": "number" },
              "weight_modifier": { "type": "number" },
              "total": { "type": "number" },
              "is_override": { "type": "boolean" }
            }
          },
          "quantity": { "type": "integer", "description": "Stock quantity" },
          "weight": { "type": "number", "description": "Weight in kg" },
          "image_url": { "type": "string" },
          "is_default": { "type": "boolean" },
          "status": { "type": "string", "enum": ["active", "inactive", "out_of_stock"] },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "ProductVariantInput": {
        "type": "object",
        "required": ["sku"],
        "properties": {
          "sku": { "type": "string", "maxLength": 100 },
          "barcode": { "type": "string" },
          "attributes": {
            "type": "object",
            "description": "Variant attributes with pricing",
            "properties": {
              "color": {
                "type": "object",
                "properties": {
                  "value": { "type": "string" },
                  "displayValue": { "type": "string" },
                  "type": { "type": "string", "enum": ["hex", "text", "image"] },
                  "priceImpact": { "type": "number", "default": 0 },
                  "metadata": { "type": "object" }
                }
              },
              "size": {
                "type": "object",
                "properties": {
                  "value": { "type": "string" },
                  "priceImpact": { "type": "number", "default": 0 }
                }
              },
              "capacity": {
                "type": "object",
                "properties": {
                  "value": { "type": "string" },
                  "displayValue": { "type": "string" },
                  "basePrice": { "type": "number" }
                }
              },
              "weight": {
                "type": "object",
                "properties": {
                  "value": { "type": "number" },
                  "label": { "type": "string" },
                  "priceImpact": { "type": "number", "default": 0 }
                }
              }
            }
          },
          "base_price": { "type": "number" },
          "override_price": { "type": "number", "description": "Manual price (skips calculation)" },
          "quantity": { "type": "integer", "minimum": 0, "default": 0 },
          "weight": { "type": "number" },
          "image_url": { "type": "string" },
          "is_default": { "type": "boolean", "default": false },
          "status": { "type": "string", "enum": ["active", "inactive", "out_of_stock"], "default": "active" }
        }
      },
      "ProductImage": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "product_id": { "type": "string", "format": "uuid" },
          "url": { "type": "string" },
          "is_main": { "type": "boolean" },
          "sort_order": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }]
}